rename: db-config <- dbConfig
